Primeiro commit

+++ src/main/java/com/tcc2/ellemVeigaOficial/services/ClienteService.java
@@ -4,6 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.tcc2.ellemVeigaOficial.repositories.UsuarioRepository;
+import com.tcc2.ellemVeigaOficial.tratamentoerros.ClienteNaoEncontradoException;
+import com.tcc2.ellemVeigaOficial.tratamentoerros.UsuarioNaoEncontradoException;
+
 import lombok.AllArgsConstructor;
 import java.util.List;
 import com.tcc2.ellemVeigaOficial.models.Cliente;
@@ -24,20 +27,19 @@ public class ClienteService {
     }
 
     public Cliente findById(Long id) {
-        return repository.findById(id).get();
+                return repository.findById(id)
+                .orElseThrow(() -> new ClienteNaoEncontradoException("Cliente com ID " + id + " não encontrado."));
     }
 
     public Cliente addCliente(Cliente cliente) {
-        if (cliente.getUsuario() != null && cliente.getUsuario().getId() != null) {
-            Usuario usuario = usuarioRepository.findById(cliente.getUsuario().getId())
-                    .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
-            cliente.setUsuario(usuario);
-        }
-
+        validarUsuario(cliente);
         return repository.save(cliente);
     }
 
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ClienteNaoEncontradoException("Cliente com ID " + id + " não existe.");
+        }
         repository.deleteById(id);
     }
 
@@ -45,21 +47,23 @@ public class ClienteService {
     public Cliente update(Long id, Cliente cliente) {
 
         if (!repository.existsById(id)) {
-            throw new RuntimeException("Cliente não encontrado");
+            throw new ClienteNaoEncontradoException("Cliente com ID " + id + " não encontrado.");
         }
         cliente.setId(id);
-
-        if (cliente.getUsuario() != null && cliente.getUsuario().getId() != null) {
-            Usuario usuario = usuarioRepository.findById(cliente.getUsuario().getId())
-                    .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
-            cliente.setUsuario(usuario);
-        }
-
+        validarUsuario(cliente);
         return repository.save(cliente);
     }
 
     public List<Cliente> buscarClientes(Long id, String nome) {
         return repository.buscarClientes(id, nome);
+    }
+
+    private void validarUsuario(Cliente cliente) {
+        if (cliente.getUsuario() != null && cliente.getUsuario().getId() != null) {
+            Usuario usuario = usuarioRepository.findById(cliente.getUsuario().getId())
+                    .orElseThrow(() -> new UsuarioNaoEncontradoException("Usuário com ID " + cliente.getUsuario().getId() + " não encontrado."));
+            cliente.setUsuario(usuario);
+        }
     }
 
 }
