+++ src/main/java/com/tcc2/ellemVeigaOficial/services/PedidoService.java
@@ -2,14 +2,17 @@ package com.tcc2.ellemVeigaOficial.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
 import java.util.List;
 import com.tcc2.ellemVeigaOficial.models.Pedido;
 import com.tcc2.ellemVeigaOficial.models.Usuario;
 import com.tcc2.ellemVeigaOficial.repositories.ClienteRepository;
+import com.tcc2.ellemVeigaOficial.repositories.PedidoProdutoRepository;
 import com.tcc2.ellemVeigaOficial.repositories.PedidoRepository;
 import com.tcc2.ellemVeigaOficial.repositories.UsuarioRepository;
+import com.tcc2.ellemVeigaOficial.repositories.VendaRepository;
 
 @Service
 public class PedidoService {
@@ -19,20 +22,12 @@ public class PedidoService {
     private UsuarioRepository usuarioRepository;
     @Autowired
     private ClienteRepository clienteRepository;
+    @Autowired
+    private PedidoProdutoRepository pedidoProdutoRepository;
+    @Autowired
+    private VendaRepository vendaRepository;
 
-    /*
-     * public Pedido addPedido(Pedido pedido){
-     * if (pedido.getUsuario() != null && pedido.getUsuario().getId() != null) {
-     * Usuario usuario =
-     * usuarioRepository.findById(pedido.getUsuario().getId()).orElseThrow(() -> new
-     * RuntimeException("Usuário não encontrado"));
-     * pedido.setUsuario(usuario);
-     * }
-     * 
-     * return repository.save(pedido);
-     * }
-     */
-
+    @Transactional
     public Pedido salvar(Pedido pedido) {
         if (pedido.getUsuario() != null && pedido.getUsuario().getId() != null) {
             pedido.setUsuario(usuarioRepository.findById(pedido.getUsuario().getId())
@@ -40,12 +35,20 @@ public class PedidoService {
         }
 
         if (pedido.getCliente() != null && pedido.getCliente().getId() != null) {
-            pedido.setCliente(clienteRepository.findById(pedido.getCliente().getId())
-                    .orElseThrow(() -> new IllegalArgumentException("Cliente não encontrado")));
+            Long clienteId = pedido.getCliente().getId();
+            if (clienteId > 0) {
+                pedido.setCliente(clienteRepository.findById(clienteId)
+                        .orElseThrow(() -> new IllegalArgumentException("Cliente não encontrado")));
+            } else {
+                pedido.setCliente(null); // ID inválido, ignora
+            }
+        } else {
+            pedido.setCliente(null); // Cliente ausente, ignora
         }
 
         return pedidoRepository.save(pedido);
     }
+
 
     public Pedido findById(Long id) {
         return pedidoRepository.findById(id).get();
@@ -55,7 +58,16 @@ public class PedidoService {
         return pedidoRepository.findAll();
     }
 
+    @Transactional
     public void delete(Long id) {
+        Pedido pedido = pedidoRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Pedido não encontrado"));
+
+         if (vendaRepository.existsByPedido(pedido)) {
+            throw new IllegalStateException("Não é possível excluir o pedido. Há uma venda vinculada a ele.");
+        }
+
+        pedidoProdutoRepository.deleteByPedido(pedido);
         pedidoRepository.deleteById(id);
     }
