+++ src/main/java/com/tcc2/ellemVeigaOficial/services/ProdutoService.java
@@ -5,7 +5,9 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import com.tcc2.ellemVeigaOficial.models.Produto;
 import com.tcc2.ellemVeigaOficial.models.Usuario;
+import com.tcc2.ellemVeigaOficial.repositories.PedidoProdutoRepository;
 import com.tcc2.ellemVeigaOficial.repositories.ProdutoRepository;
+import com.tcc2.ellemVeigaOficial.repositories.ProdutoVendaRepository;
 import com.tcc2.ellemVeigaOficial.repositories.UsuarioRepository;
 
 import lombok.AllArgsConstructor;
@@ -19,6 +21,10 @@ public class ProdutoService {
     private ProdutoRepository repository;
     @Autowired
     private UsuarioRepository usuarioRepository;
+    @Autowired
+    private PedidoProdutoRepository pedidoProdutoRepository;
+    @Autowired
+    private ProdutoVendaRepository produtoVendaRepository;
 
     public Produto addProduto(Produto produto){
         if (produto.getUsuario() != null && produto.getUsuario().getId() != null) {
@@ -37,6 +43,21 @@ public class ProdutoService {
     }
 
     public void delete(Long id){
+        // Verifica se o produto existe
+        Produto produto = repository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Produto não encontrado"));
+
+        // Verifica se está vinculado a ProdutoVenda
+        boolean vinculadoProdutoVenda = produtoVendaRepository.existsByProdutoId(id);
+
+        // Verifica se está vinculado a PedidoProduto
+        boolean vinculadoPedidoProduto = pedidoProdutoRepository.existsByProdutoId(id);
+
+        if (vinculadoProdutoVenda || vinculadoPedidoProduto) {
+            throw new RuntimeException("Não é possível excluir o produto, pois está vinculado a vendas ou pedidos.");
+        }
+
+        // Exclui o produto
         repository.deleteById(id);
     }
